refactor, stage 2
