+++ src/nodes/dateIntervalBars.ts
@@ -1,0 +1,9 @@
+import {
+  DateIntervalBarChart,
+  MongoAggregation
+} from '../types'
+
+export const results = ({ x, y, group, period }: DateIntervalBarChart): MongoAggregation => [
+  { $group: { _id: { [`${period}`]: { [`$${period}`]: `$${x}` }, group: `$${group}` }, value: { $sum: `$${y}` } } },
+]
+
