+++ src/nodes/fieldHasTruthyValue.ts
@@ -1,0 +1,13 @@
+import {
+  FieldHasTruthyValueFilter,
+  MongoAggregation
+} from '../types'
+
+export const filter = ({ field, checked }: FieldHasTruthyValueFilter): MongoAggregation =>
+    checked
+      ? [
+          {
+            $match: { [field]: { $ne: null } },
+          },
+        ]
+      : []
