+++ src/nodes/propExists.ts
@@ -1,0 +1,11 @@
+import {
+  PropExistsFilter,
+  MongoAggregation
+} from '../types'
+
+export const filter = ({ field, negate }: PropExistsFilter): MongoAggregation =>
+    [
+      {
+        $match: { [field]: { $exists: !negate } },
+      },
+    ]
